refactor: extract milestone report into modular components and fix data flow
- Break ReportMilestonePage (975 lines) into focused components:
  StatsGrid, StatsTable, PendingVerificationTable, ReportTabBar
- Extract useReportPageData hook for all data-fetching and state logic
- Extract normalizeProgramId utility for consistent programId handling
- Fix PendingVerificationMilestone type to match API response (milestoneUid
  instead of milestoneIndex), resolving duplicate React key errors
- Remove activeTab gate from pending query so badge count always loads
- Fix handleProgramSelect to use direct state instead of nuqs functional
  updater, preventing stale state when switching from Any to a program
