+++ utilities/queryKeys.ts
@@ -113,7 +113,7 @@ export const QUERY_KEYS = {
       ["isCommunityAdmin", communityUid, chainId, address, signer] as const,
     IS_ADMIN_BASE: ["isCommunityAdmin"] as const,
     PENDING_VERIFICATION: (communityId: string, page: number, programIds: string[]) =>
-      ["pending-verification-milestones", communityId, page, programIds] as const,
+      ["pendingVerificationMilestones", communityId, page, programIds] as const,
     PROJECT_UPDATES: (
       communityId: string,
       filter: string,
